chore(constants): move enums in config to statement
